+++ src/asm/stdout/stdout.cfg
@@ -1,0 +1,9 @@
+MEMORY {
+RAM: start = $0000, size=$8000, type = rw, fill = yes, fillval = $FF, file = %O;
+ROM: start = $8000, size=$7FFA, type = ro, fill = yes, fillval = $FF, file = %O;
+}
+SEGMENTS {
+ZEROPAGE: load=RAM, type=rw;
+DATA: load=RAM, type=rw, offset=$0200;
+CODE: load=RAM, type=rw, offset=$0400;
+}
